pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tournabot"
version = "0.1.0"
description = "Telegram bot for Smoothcomp team schedules"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot==21.7",
    "httpx==0.28.1",
    "beautifulsoup4==4.12.3",
    "python-dotenv==1.0.1",
    "pydantic==2.10.6",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
]

[project.scripts]
tournabot = "tournabot.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
